backend/requirements-flask.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
Pillow==10.1.0
pdfminer.six==20221105
sympy==1.12
huggingface-hub==0.20.0
paddleocr==2.7.0
paddlepaddle==2.5.2
numpy==1.24.3
